PANDO-21: Rename `*Repository` test classes -> `*DataSource`
